requirements.txt
# EchoTales Enhanced - Dependencies

# Core Framework
flask==3.0.0
flask-restful==0.3.10
flask-cors==4.0.0

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23

# AI/ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0

# Image Processing & Generation
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
matplotlib>=3.7.0

# Audio Processing (for future voice features)
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Text Processing & NLP
spacy>=3.7.0
nltk>=3.8.0
booknlp>=1.0.0
textblob>=0.17.0

# API Clients
requests>=2.31.0
openai>=1.3.0
google-generativeai>=0.3.0
stability-sdk>=0.8.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Configuration & Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
colorama>=0.4.6

# Web & API
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
httpx>=0.25.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0

# System & Performance
psutil>=5.9.0
memory-profiler>=0.61.0
line-profiler>=4.1.0

# Optional: GPU acceleration (uncomment if you have CUDA)
# torch-audio>=2.0.0+cu118
# torchtext>=0.15.0+cu118

# Optional: Advanced image models (heavy dependencies)
# diffusers>=0.24.0
# accelerate>=0.24.0
# xformers>=0.0.22  # For memory-efficient attention

# Windows-specific
# pywin32>=306  # For Windows-specific features